# docs/rebuild/openapi-baseline.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Tradier Bot API",
  "description": "Standalone backend for three worlds on one process: the Tradier options desk (balance, chain preview, managed positions with TP/SL, the opening-range and A/B auto-traders, the HOT scan and the options-flow board), the 36 Trade Desk that fronts it on a phone, and the Bot Station that launches, supervises and settles the vendored Kalshi bots (btc15, btc60, sports, parlay, gold15, silver15, oil15). Plus the super_research signal engines and the SPY/QQQ/SPX level watcher they read. SQLite persistence, per-user credential folders, paper-only by default.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/users": {
   "get": {
    "tags": [
     "users"
    ],
    "summary": "Get Users",
    "operationId": "getUsers",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/UserOut"
         },
         "type": "array",
         "title": "Response Getusers"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "users"
    ],
    "summary": "Create User",
    "operationId": "createUser",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/users/{user_id}": {
   "get": {
    "tags": [
     "users"
    ],
    "summary": "Get User",
    "operationId": "getUser",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "User Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "UserCreate": {
    "properties": {
     "username": {
      "type": "string",
      "maxLength": 64,
      "minLength": 1,
      "pattern": "^[A-Za-z0-9_-]{1,64}$",
      "title": "Username"
     },
     "email": {
      "anyOf": [
       {
        "type": "string",
        "format": "email"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "user_root_folder": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 1024,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "User Root Folder",
      "description": "Windows (D:\\...) or POSIX (/home/...) path to the user's secrets folder (API keys, PEM, wellness profile). Omit it and the server defaults to <TBOT_CUSTOMERS_ROOT>/<username> \u2014 clients should NOT hardcode machine paths."
     }
    },
    "type": "object",
    "required": [
     "username"
    ],
    "title": "UserCreate"
   },
   "UserOut": {
    "properties": {
     "user_id": {
      "type": "string",
      "title": "User Id"
     },
     "username": {
      "type": "string",
      "title": "Username"
     },
     "email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "user_root_folder": {
      "type": "string",
      "title": "User Root Folder"
     },
     "root_folder_exists": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Root Folder Exists"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     }
    },
    "type": "object",
    "required": [
     "user_id",
     "username",
     "email",
     "user_root_folder",
     "created_at",
     "updated_at"
    ],
    "title": "UserOut"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}